requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
